Renewal URI response is optional (brackets added)
From the paragraph explaining the certificate exchange:
```
If the server agrees to issue the certificate, then it creates the
certificate and provides it in its response. The server may also provide
a URI that can be used to renew the certificate, if it allows renewal
without re-validation.
```

The renewal URI is thus optional.