Mention IPv6 and round-robin DNS situation
I'm guessing that most Simple HTTPS and DVSNI challenge implementations will just pick the
first A record for the domain they are validating, and ignore AAAA records altogether.
The current text seems to assume "the host for the domain" is uniquely defined, which puts
an incorrect security model in the reader's head.