Clarify tokens in various challenges.
Use same language for each, which is particularly helpful for SimpleHTTP since
it prohibits '/' and '.' in tokens, which wind up used in path names.

Fixes some tokens with characters outside the URL-safe base64 set, and fixes one
token that was randomly different from the example tokens.

Fixes trailing commas in JSON.