Remove build/contribute instructions, clarify deadness of repo.
This commit replaces `draft-barnes-acme.md` with a notice that the
document lives elsewhere now.

This commit also removes CONTRIBUTING.md, SUBMITTING.md and all of the
build instructions/tooling. There's no reason to submit to this repo and
you will have to checkout an older commit from the history to find
content to build.